human_match/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-choices==1.7.1
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.0
get==2019.4.13
idna==2.9
Pillow==7.1.2
post==2019.4.13
psycopg2==2.8.5
public==2019.4.13
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
